Split db crate
